requirements.txt
alembic==1.4.2
autopep8==1.5.3
certifi==2020.6.20
chardet==3.0.4
idna==2.9
Mako==1.1.3
MarkupSafe==1.1.1
mysql-connector==2.2.9
pycodestyle==2.6.0
pyTelegramBotAPI==3.7.1
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.17
toml==0.10.1
urllib3==1.25.9
